docs: ingredient 부분 컴포넌트로 따로 빼서 정리 (#13)
